Camera Tracker V1
This area of code relates to the parts needed to get the green screen up and running. The current main issue is that the foreground elements are rendered, but the background fails to load. Also a new method for the tracker will need to be implemented. It is currently locked to 3DOF when it requires 6DOF.